% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoaudio.R
\name{quartoaudio}
\alias{quartoaudio}
\title{convert a quarto file to audio}
\usage{
quartoaudio(key, api = "openai", language = "en", ...)
}
\arguments{
\item{key}{API key}

\item{api}{TTS APIs (openAI, voiceRSS,)}
}
\description{
convert a quarto file to audio
}
