% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_credentials.R
\name{set_api_credentials}
\alias{set_api_credentials}
\title{set API credentials}
\usage{
set_api_credentials(api, ...)
}
\arguments{
\item{api}{API name (openAI, voiceRSS, play.ht)}
}
\description{
set API credentials
}
