% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_openai_credentials.R
\name{set_openai_credentials}
\alias{set_openai_credentials}
\title{set API credentials}
\usage{
set_openai_credentials(key = NULL)
}
\arguments{
\item{key}{openAI API key}
}
\description{
set API credentials
}
