% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoaudio.R
\name{quartoaudio}
\alias{quartoaudio}
\title{convert a quarto file to audio}
\usage{
quartoaudio(api, ...)
}
\arguments{
\item{api}{TTS APIs (openAI, voiceRSS, play.ht)}
}
\description{
convert a quarto file to audio
}
