% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoaudio_playht.R
\name{quartoaudio_playht}
\alias{quartoaudio_playht}
\title{Convert  a quarto file to audio using openAI API}
\usage{
quartoaudio_playht(input, user_id, secret_key, output_format = "mp3", ...)
}
\arguments{
\item{input}{The text to generate into audio}
}
\description{
Convert  a quarto file to audio using openAI API
}
